fix(TypePicker, Overlay, ColumnType): clean up based on pr review
